importer_app/data_scrape_job/requirements_docker.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Deprecated==1.2.13
fastapi==0.74.1
gunicorn==20.1.0
h11==0.13.0
httptools==0.3.0
idna==3.3
packaging==21.3
prometheus-client==0.13.1
pydantic==1.9.0
pyparsing==3.0.7
python-dotenv==0.19.2
PyYAML==6.0
redis==4.1.4
requests==2.27.1
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
watchgod==0.7
websockets==10.2
wrapt==1.13.3
